build: update vendor libraries and improve build configuration
The vendor libraries are updated to specific tags to ensure that the build is reproducible.
The FETCHCONTENT_QUIET is set to ON to reduce the verbosity of the build process.
The benchmark testing is disabled to reduce the build time.
The FIND_PACKAGE_ARGS is set to "" for vendor libraries to avoid conflicts with the system libraries.
The PYBIND11_FINDPYTHON is set to "NEW" to use the new Python finder.