refactor(math): move implementation of jacobi polynomials from .cpp to .hpp
The implementation of the jacobi polynomials was moved from the .cpp file to the .hpp file to allow for template definitions to be visible at the point of instantiation. This change is necessary because the compiler needs to know the definition of the template function when it is called, and this information is not available if the definition is in a separate compilation unit.