feat: add dockerfile and github action to build and push addlicense
This commit introduces a Dockerfile and a GitHub Actions workflow to
automate the building and pushing of the `addlicense` Docker image to
GitHub Container Registry (GHCR). The Dockerfile defines the steps to
build the image, while the GitHub Actions workflow automates the
process on pushes to the `master` branch and workflow dispatches.